Added horizontal wiping to increase stability.
Some filaments that don't bind well during initial stages of wiping, such as PVA and PETG, require the alternation of wipe direction between layers to vastly reduce the impact of holes and peeling due to lack of adhesion. This alternation between y-forward, x-forward, y-reversed, and x-reversed wiping acts to self-correct any mistakes in previous layers.

I also increased the extrusion width during the wiping of non-first layers to match the first layer's width. This was not allowing the layers to bridge over small gaps properly.

I would like to make both of these a toggle or a setting, but do not know my way around the rest of the code yet.