Undo ineffective bugfix
When the wipe tower is thin and long, which is sometimes useful for bridging when some materials peel on the first few wipes, the toolchange_Unload initial position is positioned over the edge of the wipe tower. This behavior doesn't seem to be intended. I put in a very temporary fix that produced the inverse issue (starting too far inside the perimeter) when the wipe tower is in the default shape, so I undid it.