Fix permission handling in copy_file, fixes #2521
Previously we used to try to set permission on all saves regardless of
whether we needed to. This broke saving on filesystems that don't
support chmod'ing the files (or probably cases where you could write the
file but not chmod it). This changeset instead just tries saving, and
only if/when it fails due to permissions, it tries changing the
permissions.