Add setting for controlling density of first layer support lines.
This adds a setting under "Print Settings" for controlling the density
of first layer support lines, which previously was hard-coded to 0.5.
Changing it to 1.0 yields a solid first layer and making it less than
0.5 yields even sparser support lines resulting in fewer corners.

Frankly the first layer adhesion of the textured metal sheets is not
great and really struggles with small features on the print bed.  This
is exacerbated by the first layer support pattern with its very sharp
angles.  This change gives the user the option to increase the density
to get a solid support layer for better adhesion, or to decrease the
density and reduce the number of sharp corners.

Partially addresses issue #2101.