Implement the "Uniform" infill mode
"Uniform" is essentially rectilinear infill where the filling direction
for the layer is oriented and fixed in a consistent ltr or rtl pattern.

"Uniform" is mostly useful as a top layer to improve visual appearance
by reducing the different reflections caused by a randomly changing
filling direction. The constrained orientation might overall increase
the number of travels, but it's generally very close to Rectilinear.

We implement Uniform by adding a new virtual method in FillRectilinear2
to control the new ordering behavior. While filling, the visiting order
of the intersections is tweaked to favor an even/odd rule matching the
current vertical intersection. Similarly, when connecting between
vertical lines, only one direction is allowed.