Storing relative paths in 3mf.
This is to fix #5540

Paths never make sense to store as full paths as discussed
in #5540 - it produces files that are not useful on other
systems than the place where the 3mf is creates.
But just the security aspect alone (leaking filesystem structure
of the creating system is important).

This patch changes the output to be relative to the 3mf file generated.

Signed-off-by: Henner Zeller <h.zeller@acm.org>