feat: add option for minimum filament use on a single-extruder print
This is useful for the Mosaic Palette (especially when used with P2PP),
as the Palette can only splice segments of at least a given length
reliably.

Adds an option on the printer settings to set that length, and, when
filled, ensures that at least that amount of filament will be used
before switching tools, likely by increasing wipe length. This allows
some savings over the previously-recommended P2PP settings of
unilaterally increasing wipe volumes, as fixed wipe volumes would rarely
be required on every use of a given filament.