Bonjour: per-interface UDP sockets
If there is more than one interface available on a system, datagrams are
sent using a default interface (as specified both by BSD Sockets and
Winsock).

The printer and machine's default gateway are typically parts of the same
network, in which case datagrams broadcasted through a default interface
will reach the printer. However, in a scenario the tiniest bit more
complicated than this, the comfort of automatic service discovery gets
lost.

Thereby I suggest going the extra mile and sending the mDNS query via
every interface that's assigned one of the IPs to which machine's
hostname is resolved.