WIP: Port superslicer calibration to prusaslicer
Fixes #6168

This ports superslicer's calibration menu to prusaslicer.
The approach was generally:
1. Copy files from superslicer to prusaslicer
2. Attempt to compile
3. Fix errors

I had to make a few changes in other files because superslicer has
different config options than prusaslicer.

What's working is opening the calibration menu and generating some of
the prints.

However when prusaslicer encounters a config key it doesn't know about
there's a segfault.

Some of the settings are needed to generate specific tests, so either
need to implement the new settings or generate the test another way.