Add commandline option for specifying the build tool, move vswhere search to call
Added the -t -TOOL command line option to specify which build environment to use. Choices are auto, visual, and msbuild
The vswhere.exe search had too be moved to a call, to make it easier to specify where it should be used. It now must be used after parsing the command line options. Need to verify that this didn't break anything (so far it builds successfully).