format/AnycubicSLA: Add missing fields
The AnycubicSLA format description is derived from reverse engineering of
the PhotonWorkshop output files.

While the initial Photon devices had their binary headers with version 1
we have seen newer models with additional versions. Namely 515, 516 and 517.
We are adding the version handling to the AnycubicSLA exporter to prepare
for future version handling.
Some fields were missing for Version 1 which are added.

Signed-off-by: Mimoja <git@mimoja.de>