Improve overhang perimeter printability with large extrusion widths
====================================================================

What for?

When printing with relatively slower printers, such as MK3 (no input shaper), no substantial time
savings can be achieved by increasing the speed without substantial loss of quality.

For a fixed nozzle size, reduction of print times can come from the remaining parameters defining
the flow rate within the volumetric flow rate of the extruder (mm^3/s), i.e. extrusion width and
layer height.

When a part has a substantial amount of relatively thin walls, increasing the extrusion width may
lead to reducing the number of perimeters (due to wider lesser amount of perimeters) while keeping
the speed constant.

What is the problem when using larger extrusion widths?

A MK3 can make reasonably good parts even at 200% extrusion width and even beyond. For a 0.4mm
nozzle, those are extrusion widths of 0.8-1 mm.

This video shows some tests at different extrusion widths:
https://www.youtube.com/watch?v=9YaJ0wSKKHA

However, problems arise with overhang perimeters (perimeters used for bridging), as wider perimeters
weight more, are spouted out of the relatively small hole of the nozzle energetically and sag. These
are also a very poor support for the next layers.

Does thick bridges affect the outcome?

This problem arises both, with the "Thick bridges" enabled and disabled.

With it enabled, the relatively thick perimeters (EW=0.9mm) are reduced to the nozzle width (0.4mm)
over the overhang area (and are printed properly). However, the spacing between bridging perimeters
remains that corresponding to an EW of 0.9mm, this provides a very poor support of the next layers,
with often the material of the next layer being forced through or over the holes between wires (to
the broad spacing).

With it disabled, the overhang perimeters are too thick themselves and sag, making it a poor support
for the next layer.

So what is the parameter that may be tweaked?

The separation between overhanging perimeters can be reduced, so that it corresponds to the spacing
when the EW is 0.4mm. Then the EW of these wires need to be reduced accordingly.

How can this be achieved?

One could provide more wires just over the overhang area. This would lead to a solution where the
number of "perimeters" is different over the overhang area and the non-overhang area, within a single
layer region. This appears to drive Slic3r crazy, as it wants full loops over a single layer region.

So what is the solution you propose?

The solution is to reduce the extrusion width of the whole region comprising the overhang area, so that
the perimeter generator generates more lighter fine spaced perimeters, which are a better support for
the next layers.

There appears to be no way to know whether a layer will have overhangs before creating the extrusions for
a given layer. The solution, which is here only implemented for Arachne, is essentially to run Arachne and
if there are overhangs, Arachne is interrupted and re-run with a reduced extrusion width.

Is this a general mechanism?

No, this mechanism is only enabled if the EW is higher than 1.2*the_nozzle_diameter, so not to interfere with
traditional ways of slicing (where it would not have any effect anyway, as EW are relatively speaking not "large").
Also, the solution is not applied for every single overhanging perimeter, as for relatively short perimeters, the
default values are just fine. A new configuration is provided in "Advanced", just after "Bridge flow ratio". For
overhangs lower than that value, the traditional slicing is used. When that threshold is surpassed, then the new
mechanism comes into play.

What is the role of bridge flow ratio?

In case that the overhang perimeters are not enough to cover the whole width, solid infill is used to
fill in the gap. If the extrusion width for the solid infil is also large (e.g. 200% of nozzle diameter),
the problem with the perimeters is replicated for the bridging infill (it is too thick, unless thick bridges
is selected). However, the solid infill is scaled with the bridge flow ratio (advanced print settings). It
is possible to use the bridge flow ratio to scale down the infill solid extrusion during bridging and avoid
the sagging. For example, if the Solid infil is at 200%, it is possible to configure the bridge flow ratio
to 0.5, so that the extrusion width of solid infil during bridging is made 0.4mm (when using a 0.4mm nozzle).
That is the case when "thick bridges" (in perimeters) is disabled. If it is enabled, the solid infil still
scales the diameter of the nozzle (so it can be used to reduce the infil, but it is not scaling the solid
infil setting in advanced print settings, but the diameter nozzle). See LayerRegion::bridging_flow for more
detailed information.