Fix skirt separation math.
In the code as-is, the half-width of the skirt extrusion is
added to the user-specified distance.  However, the user-specified
distance is scaled to internal units while the half-width is not,
resulting in math like "250000 + 0.125" for a skirt extrusion of
0.5mm

The fix is to include the half-width inside the scaling function
so that it is also scaled, resulting in math like "250000 + 125000".

The user-visible result is that if you have no brim and set the
skirt spacing to zero, the skirt will actually touch the model
and not be offset by half an extrusion width.